Deprecate io.vertx.ext.auth.User
