+++ vertx-auth-common/src/main/java/io/vertx/ext/auth/authentication/AuthenticationProvider.java
@@ -16,6 +16,7 @@
 
 package io.vertx.ext.auth.authentication;
 
+import io.vertx.codegen.annotations.GenIgnore;
 import io.vertx.codegen.annotations.Fluent;
 import io.vertx.codegen.annotations.GenIgnore;
 import io.vertx.codegen.annotations.VertxGen;
