+++ vertx-auth-common/src/main/java/io/vertx/ext/auth/authorization/impl/AuthorizationContextImpl.java
@@ -15,8 +15,8 @@ package io.vertx.ext.auth.authorization.impl;
 import java.util.Objects;
 
 import io.vertx.core.MultiMap;
+import io.vertx.ext.auth.user.User;
 import io.vertx.ext.auth.authorization.AuthorizationContext;
-import io.vertx.ext.auth.User;
 
 public class AuthorizationContextImpl implements AuthorizationContext {
