+++ vertx-auth-common/src/main/java/io/vertx/ext/auth/impl/UserImpl.java
@@ -104,7 +104,7 @@ public class UserImpl implements User, ClusterSerializable {
   }
 
   @Override
-  public User merge(User other) {
+  public User merge(io.vertx.ext.auth.user.User other) {
     if (other == null) {
       return this;
     }
